requirements.txt
feedparser
beautifulsoup4
google-generativeai
google-cloud-bigquery
pandas
python-dotenv
